emergency_storage/requirements.txt
requests~=2.31.0
Flask~=2.3.2
PyMySQL~=1.1.0
SQLAlchemy~=2.0.17
python-dotenv~=1.0.0